# api/openapi.yaml
...
openapi: 3.2.0
info:
  title: Layover Tower API
  version: 0.2.0
  summary: Control surface for a local-first agent factory.
  description: |
    The UI is purely a client of this API, so this document bounds what the UI can ever do.

    Everything is local-first and unauthenticated by default: the Tower binds to loopback. Do not
    expose it without putting something in front of it.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://127.0.0.1:7878
    description: The default local Tower.
...
paths:
...
  /flights:
    post:
      operationId: sendFlight
      summary: Start work by sending the first flight of a new itinerary.
      description: |
        Give either a `pipeline` or a `to`. A pipeline is the normal way in: it names the entry
        agent and declares which flags may be set. A bare `to` sends to an agent marked
        `entry = true` and accepts no flags.
      tags: [work]
      requestBody:
        required: true
        description: What to run, and with which flags.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendFlightRequest"
      responses:
        "202":
          description: The flight was accepted and a run is being started.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlightAccepted"
        "400":
          description: The request named an unknown pipeline, agent or flag.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: A Ground Stop is engaged, so no new work may start.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /runs:
    get:
      operationId: listRuns
      summary: Runs, live and historical.
      tags: [work]
      parameters:
        - name: status
          in: query
          required: false
          description: Return only runs in this state.
          schema:
            $ref: "#/components/schemas/RunStatus"
        - name: itinerary_id
          in: query
          required: false
          description: Return only runs belonging to this itinerary.
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: Maximum number of runs to return, newest first.
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: Matching runs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunList"
  /runs/{run_id}:
    get:
      operationId: getRun
      summary: One run, including how it ended.
      tags: [work]
      parameters:
        - name: run_id
          in: path
          required: true
          description: Identifier of the run to fetch.
          schema:
            type: string
      responses:
        "200":
          description: The run.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "404":
          description: No such run.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /runs/{run_id}/stream:
    get:
      operationId: streamRun
      summary: Live output from a run, as server-sent events.
      tags: [work]
      parameters:
        - name: run_id
          in: path
          required: true
          description: Identifier of the run to stream.
          schema:
            type: string
      responses:
        "200":
          description: A stream of output lines, ending when the run exits.
          content:
            text/event-stream:
              schema:
                type: string
                description: One output line per event.
        "404":
          description: No such run.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
...
    SendFlightRequest:
      type: object
      description: |
        Exactly one of `pipeline` and `to` must be given, and flags are only accepted alongside a
        pipeline, because a pipeline is what declares them.

        Neither is marked required here because JSON Schema cannot state "exactly one of these"
        in a way this generator supports. The Tower enforces it and answers `400` — so a client
        that satisfies this schema can still be rejected.
      required: [body]
      properties:
        pipeline:
          type: [string, "null"]
          description: Name of the pipeline to start.
        to:
          type: [string, "null"]
          description: Name of an `entry = true` agent to send to directly.
        body:
          type: string
          description: What the receiving agent is being asked to do.
        flags:
          type: object
          description: Flag overrides, keyed by flag name. Unset flags take their declared default.
          additionalProperties:
            type: boolean
            description: Whether the flag is on.
    FlightAccepted:
      type: object
      description: Confirmation that work has been started.
      required: [flight_id, itinerary_id, to]
      properties:
        flight_id:
          type: string
          description: Identifier of the flight that was created.
        itinerary_id:
          type: string
          description: Identifier of the new chain, which carries Hops and Fuel.
        to:
          type: string
          description: The agent the flight was addressed to.
    RunStatus:
      type: string
      description: |
        How a run ended, or that it has not. `stalled` is deliberately distinct from `failed`: a
        factory that quietly parks work forever is worse than one that crashes, so it has to be
        visible as its own outcome.
      enum: [running, succeeded, failed, stalled, cancelled]
    Run:
      type: object
      description: One supervised CLI execution.
      required: [run_id, itinerary_id, agent, status, started_at]
      properties:
        run_id:
          type: string
          description: Identifier of this run.
        itinerary_id:
          type: string
          description: The chain this run belongs to.
        agent:
          type: string
          description: Which agent was run.
        status:
          $ref: "#/components/schemas/RunStatus"
          description: How it ended, or that it has not.
        started_at:
          type: string
          format: date-time
          description: When the process was spawned.
        finished_at:
          type: [string, "null"]
          format: date-time
          description: When the process exited; null while it is still running.
        exit_code:
          type: [integer, "null"]
          format: int32
          description: Process exit code, when there was one.
        cost_usd:
          type: [number, "null"]
          format: double
          description: Null when the runner reported no cost, which the Tower logs loudly.
        hops_remaining:
          type: [integer, "null"]
          format: int32
          description: Hops left when this run was started, mirrored for display only.
    RunList:
      type: object
      description: Matching runs, newest first.
      required: [runs]
      properties:
        runs:
          type: array
          description: The runs.
          items:
            $ref: "#/components/schemas/Run"
...
    Problem:
      type: object
      description: An error, shaped after RFC 9457.
      required: [title, status]
      properties:
        title:
          type: string
          description: A short, human-readable summary of the problem.
        status:
          type: integer
          format: int32
          description: The HTTP status code, repeated in the body for clients that lose it.
        detail:
          type: [string, "null"]
          description: An explanation specific to this occurrence.