# crates/layover-http/src/generated.rs
//! The Layover HTTP surface, generated from `api/openapi.yaml`.
//!
//! @generated by `cargo xtask generate-api`. Do not edit by hand: `cargo xtask verify`
//! regenerates this file and fails if the result differs, so an edit here is reverted
//! rather than kept. Change the specification instead.
//!
//! Source: Layover Tower API v0.2.0

#![allow(clippy::too_many_lines)]

use axum::response::IntoResponse as _;
use serde::{Deserialize, Serialize};

use crate::EventStream;

/// Whether an agent writes to the shared workspace or gets a snapshot.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
pub enum Access {
    /// `read-only`
    #[serde(rename = "read-only")]
    ReadOnly,
    /// `read-write`
    #[serde(rename = "read-write")]
    ReadWrite,
}

/// One configured agent.
#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
pub struct Agent {
    /// Whether the agent writes to the shared workspace.
    pub access: Access,
    /// One line saying what the agent is.
    #[serde(default)]
    pub description: Option<String>,
    /// Whether a human may send flights straight to this agent.
    pub entry: bool,
    /// Model identifier passed to the runner.
    #[serde(default)]
    pub model: Option<String>,
    /// The key the agent is declared under, and what routes refer to.
    pub name: String,
    /// A longer statement of when to route work here.
    #[serde(default)]
    pub purpose: Option<String>,
    /// Whether the agent is pinned resident rather than transient.
    pub resident: bool,
    /// Which runner invokes this agent's CLI.
    #[serde(default)]
    pub runner: Option<String>,
}

/// The factory's agents and the edges between them.
#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
pub struct AgentList {
    /// Every configured agent.
    pub agents: Vec<Agent>,
    /// The route map.
    pub routes: Vec<Route>,
}

/// A boolean parameter a pipeline accepts at trigger time.
#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
pub struct Flag {
    /// Value used when the trigger does not set the flag.
    pub default: bool,
    /// What turning this flag on actually does.
    #[serde(default)]
    pub description: Option<String>,
    /// The name a prompt tests with `@include(name)`.
    pub name: String,
}

/// Confirmation that work has been started.
#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
pub struct FlightAccepted {
    /// Identifier of the flight that was created.
    pub flight_id: String,
    /// Identifier of the new chain, which carries Hops and Fuel.
    pub itinerary_id: String,
    /// The agent the flight was addressed to.
    pub to: String,
}

/// Whether everything is halted.
#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
pub struct GroundStop {
    /// True when no new work may start.
    pub engaged: bool,
    /// When the Ground Stop was engaged.
    #[serde(default)]
    pub since: Option<String>,
}

/// Whether the Tower is up and whether it is allowed to do anything.
#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
pub struct Health {
    /// True when a Ground Stop is engaged and no new work may start.
    pub ground_stop: bool,
    /// Always `ok`; failure is signalled by not answering at all.
    pub status: Status,
    /// The Tower's own version.
    pub version: String,
}

/// The condition under which a rendezvous barrier releases.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
pub enum Join {
    /// `all`
    #[serde(rename = "all")]
    All,
    /// `any`
    #[serde(rename = "any")]
    Any,
}

/// A named entry point into the mesh.
#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
pub struct Pipeline {
    /// One line saying what this pipeline is for.
    #[serde(default)]
    pub description: Option<String>,
    /// The agent that receives the first flight.
    pub entry: String,
    /// The boolean parameters this pipeline accepts.
    pub flags: Vec<Flag>,
    /// The key the pipeline is declared under.
    pub name: String,
    /// What starts it.
    pub trigger: Trigger,
}

/// Every declared pipeline.
#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
pub struct PipelineList {
    /// The pipelines.
    pub pipelines: Vec<Pipeline>,
}

/// An error, shaped after RFC 9457.
#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
pub struct Problem {
    /// An explanation specific to this occurrence.
    #[serde(default)]
    pub detail: Option<String>,
    /// The HTTP status code, repeated in the body for clients that lose it.
    pub status: i32,
    /// A short, human-readable summary of the problem.
    pub title: String,
}

/// One entry of the route map, possibly expanding to several edges.
#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
pub struct Route {
    /// Sending agents.
    pub from: Vec<String>,
    /// Set when flights on this edge are parked at a rendezvous barrier.
    #[serde(default)]
    pub join: Option<Join>,
    /// Backstop for a barrier that never completes.
    #[serde(default)]
    pub timeout_sec: Option<i64>,
    /// Receiving agents.
    pub to: Vec<String>,
}

/// One supervised CLI execution.
#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
pub struct Run {
    /// Which agent was run.
    pub agent: String,
    /// Null when the runner reported no cost, which the Tower logs loudly.
    #[serde(default)]
    pub cost_usd: Option<f64>,
    /// Process exit code, when there was one.
    #[serde(default)]
    pub exit_code: Option<i32>,
    /// When the process exited; null while it is still running.
    #[serde(default)]
    pub finished_at: Option<String>,
    /// Hops left when this run was started, mirrored for display only.
    #[serde(default)]
    pub hops_remaining: Option<i32>,
    /// The chain this run belongs to.
    pub itinerary_id: String,
    /// Identifier of this run.
    pub run_id: String,
    /// When the process was spawned.
    pub started_at: String,
    /// How it ended, or that it has not.
    pub status: RunStatus,
}

/// Matching runs, newest first.
#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
pub struct RunList {
    /// The runs.
    pub runs: Vec<Run>,
}

/// How a run ended, or that it has not. `stalled` is deliberately distinct from `failed`: a
/// factory that quietly parks work forever is worse than one that crashes, so it has to be
/// visible as its own outcome.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
pub enum RunStatus {
    /// `running`
    #[serde(rename = "running")]
    Running,
    /// `succeeded`
    #[serde(rename = "succeeded")]
    Succeeded,
    /// `failed`
    #[serde(rename = "failed")]
    Failed,
    /// `stalled`
    #[serde(rename = "stalled")]
    Stalled,
    /// `cancelled`
    #[serde(rename = "cancelled")]
    Cancelled,
}

/// Exactly one of `pipeline` and `to` must be given, and flags are only accepted alongside a
/// pipeline, because a pipeline is what declares them.
///
/// Neither is marked required here because JSON Schema cannot state "exactly one of these"
/// in a way this generator supports. The Tower enforces it and answers `400` — so a client
/// that satisfies this schema can still be rejected.
#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
pub struct SendFlightRequest {
    /// What the receiving agent is being asked to do.
    pub body: String,
    /// Flag overrides, keyed by flag name. Unset flags take their declared default.
    #[serde(default)]
    pub flags: Option<std::collections::BTreeMap<String, bool>>,
    /// Name of the pipeline to start.
    #[serde(default)]
    pub pipeline: Option<String>,
    /// Name of an `entry = true` agent to send to directly.
    #[serde(default)]
    pub to: Option<String>,
}

/// Liveness of the Tower itself.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
pub enum Status {
    /// `ok`
    #[serde(rename = "ok")]
    Ok,
}

/// What starts a pipeline.
#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
pub struct Trigger {
    /// Set when the pipeline fires on a five-field cron expression.
    #[serde(default)]
    pub cron: Option<String>,
    /// Set when the pipeline fires on a fixed interval.
    #[serde(default)]
    pub every_seconds: Option<i64>,
    /// Whether a human or a clock starts this pipeline.
    pub kind: TriggerKind,
}

/// Whether a human or a clock starts a pipeline.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
pub enum TriggerKind {
    /// `manual`
    #[serde(rename = "manual")]
    Manual,
    /// `scheduled`
    #[serde(rename = "scheduled")]
    Scheduled,
}

/// query parameters for `listRuns`.
#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
pub struct ListRunsQuery {
    /// Return only runs in this state.
    #[serde(default)]
    pub status: Option<RunStatus>,
    /// Return only runs belonging to this itinerary.
    #[serde(default)]
    pub itinerary_id: Option<String>,
    /// Maximum number of runs to return, newest first.
    #[serde(default)]
    pub limit: Option<i32>,
}

/// path parameters for `getRun`.
#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
pub struct GetRunPath {
    /// Identifier of the run to fetch.
    pub run_id: String,
}

/// path parameters for `streamRun`.
#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
pub struct StreamRunPath {
    /// Identifier of the run to stream.
    pub run_id: String,
}

/// Everything the Tower must implement to serve this API.
///
/// One method per `operationId`. Adding an operation to `api/openapi.yaml` adds a method
/// here, so an unimplemented endpoint is a compile error rather than a 404 discovered in
/// production.
pub trait Api: Send + Sync + 'static {
    /// Every configured agent and the route map between them.
    ///
    /// This is the UI's graph view, and it is also how a human answers "what can talk to what"
    /// without reading the TOML.
    ///
    /// `GET /agents`
    fn list_agents(&self) -> impl core::future::Future<Output = Result<AgentList, Problem>> + Send;
    /// Start work by sending the first flight of a new itinerary.
    ///
    /// Give either a `pipeline` or a `to`. A pipeline is the normal way in: it names the entry
    /// agent and declares which flags may be set. A bare `to` sends to an agent marked
    /// `entry = true` and accepts no flags.
    ///
    /// `POST /flights`
    fn send_flight(
        &self,
        body: SendFlightRequest,
    ) -> impl core::future::Future<Output = Result<FlightAccepted, Problem>> + Send;
    /// Halt everything.
    ///
    /// Ground Stop is a file on disk rather than in-memory state, so it survives a Tower crash
    /// and can be set by hand when nothing else is responding.
    ///
    /// `POST /ground-stop`
    fn engage_ground_stop(
        &self,
    ) -> impl core::future::Future<Output = Result<GroundStop, Problem>> + Send;
    /// Resume.
    ///
    /// `DELETE /ground-stop`
    fn release_ground_stop(
        &self,
    ) -> impl core::future::Future<Output = Result<GroundStop, Problem>> + Send;
    /// Liveness, version and whether a Ground Stop is engaged.
    ///
    /// `GET /health`
    fn get_health(&self) -> impl core::future::Future<Output = Result<Health, Problem>> + Send;
    /// Every declared pipeline, its trigger and the flags it accepts.
    ///
    /// `GET /pipelines`
    fn list_pipelines(
        &self,
    ) -> impl core::future::Future<Output = Result<PipelineList, Problem>> + Send;
    /// Runs, live and historical.
    ///
    /// `GET /runs`
    fn list_runs(
        &self,
        query: ListRunsQuery,
    ) -> impl core::future::Future<Output = Result<RunList, Problem>> + Send;
    /// One run, including how it ended.
    ///
    /// `GET /runs/{run_id}`
    fn get_run(
        &self,
        path: GetRunPath,
    ) -> impl core::future::Future<Output = Result<Run, Problem>> + Send;
    /// Live output from a run, as server-sent events.
    ///
    /// `GET /runs/{run_id}/stream`
    fn stream_run(
        &self,
        path: StreamRunPath,
    ) -> impl core::future::Future<Output = Result<EventStream, Problem>> + Send;
}

/// Builds the axum router for this API.
///
/// Routes come straight from the specification, so a path can only exist here if it
/// exists there.
pub fn router<A: Api>(api: std::sync::Arc<A>) -> axum::Router {
    axum::Router::new()
        .route("/agents", axum::routing::get(handle_list_agents::<A>))
        .route("/flights", axum::routing::post(handle_send_flight::<A>))
        .route(
            "/ground-stop",
            axum::routing::post(handle_engage_ground_stop::<A>)
                .delete(handle_release_ground_stop::<A>),
        )
        .route("/health", axum::routing::get(handle_get_health::<A>))
        .route("/pipelines", axum::routing::get(handle_list_pipelines::<A>))
        .route("/runs", axum::routing::get(handle_list_runs::<A>))
        .route("/runs/{run_id}", axum::routing::get(handle_get_run::<A>))
        .route(
            "/runs/{run_id}/stream",
            axum::routing::get(handle_stream_run::<A>),
        )
        .with_state(api)
}

async fn handle_list_agents<A: Api>(
    axum::extract::State(api): axum::extract::State<std::sync::Arc<A>>,
) -> axum::response::Response {
    match api.list_agents().await {
        Ok(value) => (axum::http::StatusCode::OK, axum::Json(value)).into_response(),
        Err(problem) => problem.into_response(),
    }
}

async fn handle_send_flight<A: Api>(
    axum::extract::State(api): axum::extract::State<std::sync::Arc<A>>,
    axum::Json(body): axum::Json<SendFlightRequest>,
) -> axum::response::Response {
    match api.send_flight(body).await {
        Ok(value) => (axum::http::StatusCode::ACCEPTED, axum::Json(value)).into_response(),
        Err(problem) => problem.into_response(),
    }
}

async fn handle_engage_ground_stop<A: Api>(
    axum::extract::State(api): axum::extract::State<std::sync::Arc<A>>,
) -> axum::response::Response {
    match api.engage_ground_stop().await {
        Ok(value) => (axum::http::StatusCode::OK, axum::Json(value)).into_response(),
        Err(problem) => problem.into_response(),
    }
}

async fn handle_release_ground_stop<A: Api>(
    axum::extract::State(api): axum::extract::State<std::sync::Arc<A>>,
) -> axum::response::Response {
    match api.release_ground_stop().await {
        Ok(value) => (axum::http::StatusCode::OK, axum::Json(value)).into_response(),
        Err(problem) => problem.into_response(),
    }
}

async fn handle_get_health<A: Api>(
    axum::extract::State(api): axum::extract::State<std::sync::Arc<A>>,
) -> axum::response::Response {
    match api.get_health().await {
        Ok(value) => (axum::http::StatusCode::OK, axum::Json(value)).into_response(),
        Err(problem) => problem.into_response(),
    }
}

async fn handle_list_pipelines<A: Api>(
    axum::extract::State(api): axum::extract::State<std::sync::Arc<A>>,
) -> axum::response::Response {
    match api.list_pipelines().await {
        Ok(value) => (axum::http::StatusCode::OK, axum::Json(value)).into_response(),
        Err(problem) => problem.into_response(),
    }
}

async fn handle_list_runs<A: Api>(
    axum::extract::State(api): axum::extract::State<std::sync::Arc<A>>,
    axum::extract::Query(query): axum::extract::Query<ListRunsQuery>,
) -> axum::response::Response {
    match api.list_runs(query).await {
        Ok(value) => (axum::http::StatusCode::OK, axum::Json(value)).into_response(),
        Err(problem) => problem.into_response(),
    }
}

async fn handle_get_run<A: Api>(
    axum::extract::State(api): axum::extract::State<std::sync::Arc<A>>,
    axum::extract::Path(path): axum::extract::Path<GetRunPath>,
) -> axum::response::Response {
    match api.get_run(path).await {
        Ok(value) => (axum::http::StatusCode::OK, axum::Json(value)).into_response(),
        Err(problem) => problem.into_response(),
    }
}

async fn handle_stream_run<A: Api>(
    axum::extract::State(api): axum::extract::State<std::sync::Arc<A>>,
    axum::extract::Path(path): axum::extract::Path<StreamRunPath>,
) -> axum::response::Response {
    match api.stream_run(path).await {
        Ok(stream) => stream.into_response(),
        Err(problem) => problem.into_response(),
    }
}

/// Every operation the specification declares, as (method, path, operationId).
///
/// Exposed so tests can assert the router and the specification agree.
pub const OPERATIONS: [(&str, &str, &str); 9] = [
    ("GET", "/agents", "listAgents"),
    ("POST", "/flights", "sendFlight"),
    ("POST", "/ground-stop", "engageGroundStop"),
    ("DELETE", "/ground-stop", "releaseGroundStop"),
    ("GET", "/health", "getHealth"),
    ("GET", "/pipelines", "listPipelines"),
    ("GET", "/runs", "listRuns"),
    ("GET", "/runs/{run_id}", "getRun"),
    ("GET", "/runs/{run_id}/stream", "streamRun"),
];